vlm-inference-test/requirements.txt
transformers>=4.44.0
torch
Pillow
pyyaml
loguru
torchvision
accelerate
jinja2>=3.1.0